Bring back functions that are used in rancher
